requirements.txt
beautifulsoup4==4.6.3
bs4==0.0.1
certifi==2018.8.24
chardet==3.0.4
Django==2.1.1
django-filter==2.0.0
django-jalali-date==0.2.9
django-model-report==0.2.1
django-report-builder==6.2.2
django-rest-auth==0.9.3
djangorestframework==3.8.2
et-xmlfile==1.0.1
ez-setup==0.9
idna==2.7
jdatetime==3.0.2
jdcal==1.4
Markdown==2.6.11
openpyxl==2.5.6
Pillow==5.2.0
psycopg2==2.7.5
python-dateutil==2.7.3
pytz==2018.5
requests==2.19.1
restaurant==0.1.0
six==1.11.0
urllib3==1.23
